refactored limit order book library
